+++ packages/components/src/Radio/Radio.js
@@ -17,7 +17,6 @@ const RadioWrapper = styled(tag.div)`
 `;
 
 const RadioInput = styled.input`
-  appearance: none;
   opacity: 0;
   position: absolute;
   z-index: 0;
@@ -37,7 +36,7 @@ const RadioIcon = styled(tag.div)`
   width: ${rem('20px')};
   background-color: ${themeGet('colors.white')};
   border: ${themeGet('borders.2')};
-  border-color: ${themeGet('colors.grey.2')};
+  border-color: ${themeGet('colors.greys.alto')};
   border-radius: 50%;
   z-index: 1;
 
@@ -45,7 +44,7 @@ const RadioIcon = styled(tag.div)`
   input:focus + &,
   input:hover + & {
     border-color: ${themeGet('colors.brand.secondary')};
-    transition: ${themeGet('transitions.default')}
+    transition: ${themeGet('transitions.default')};
   }
 
   input:checked + & {
@@ -58,13 +57,13 @@ const RadioIcon = styled(tag.div)`
       height: ${themeGet('space.2')};
       width: ${themeGet('space.2')};
       border-radius: 50%;
-      background-color: ${themeGet('colors.grey.0')};
+      background-color: ${themeGet('colors.greys.charcoal')};
     }
   }
 
   input:disabled + & {
-    background-color: ${themeGet('colors.grey.2')};
-    border-color: ${themeGet('colors.grey.2')};
+    background-color: ${themeGet('colors.greys.alto')};
+    border-color: ${themeGet('colors.greys.alto')};
   }
 `;