+++ packages/test-utils/src/shallowWithTheme.js
@@ -3,6 +3,9 @@ import { shallow } from 'enzyme';
 import { ThemeProvider } from 'styled-components';
 
 export default (tree, theme) => {
-  const context = shallow(<ThemeProvider theme={theme} />).instance().getChildContext();
+  const context = shallow(<ThemeProvider theme={theme} />)
+    .instance()
+    .getChildContext();
+
   return shallow(tree, { context });
 };