+++ packages/logos/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="0.50.0"></a>
+# [0.50.0](https://github.com/hooroo/roo-ui/compare/v0.49.0...v0.50.0) (2018-08-01)
+
+
+### Features
+
+* **logo:** add Qantas Business Rewards logo ([#134](https://github.com/hooroo/roo-ui/issues/134)) ([55003de](https://github.com/hooroo/roo-ui/commit/55003de))
+
+
+
+
 <a name="0.41.0"></a>
 # [0.41.0](https://github.com/hooroo/roo-ui/compare/v0.40.2...v0.41.0) (2018-07-09)